Render events the same on their event page as on the map
Pull out rendering of points as circles into a new component,
EventFeature, that can be used on both the EventMap and Event
components.
